requirements.txt
paho-mqtt==1.5.0
requests==2.21.0
pyicloud==1.0.0